+++ packages/analytics-js/src/components/configManager/constants.ts
@@ -3,6 +3,7 @@ import type { PluginName } from '@rudderstack/analytics-js-common/types/PluginsM
 const DEFAULT_STORAGE_ENCRYPTION_VERSION = 'v3';
 
 export const ConsentManagersToPluginNameMap: Record<string, PluginName> = {
+  iubenda: 'IubendaConsentManager',
   oneTrust: 'OneTrustConsentManager',
   ketch: 'KetchConsentManager',
   custom: 'CustomConsentManager',