+++ packages/analytics-js/src/components/pluginsManager/defaultPluginsList.ts
@@ -10,6 +10,7 @@ const defaultOptionalPluginsList: PluginName[] = [
   'ErrorReporting',
   'ExternalAnonymousId',
   'GoogleLinker',
+  'IubendaConsentManager',
   'KetchConsentManager',
   'NativeDestinationQueue',
   'OneTrustConsentManager',