+++ packages/analytics-js/src/components/pluginsManager/pluginNames.ts
@@ -15,6 +15,7 @@ const pluginNamesList: PluginName[] = [
   'ErrorReporting',
   'ExternalAnonymousId',
   'GoogleLinker',
+  'IubendaConsentManager',
   'KetchConsentManager',
   'NativeDestinationQueue',
   'OneTrustConsentManager',