+++ packages/analytics-js/src/components/utilities/loadOptions.ts
@@ -33,6 +33,8 @@ const normalizeLoadOptions = (
 
   normalizedLoadOpts.secureCookie = normalizedLoadOpts.secureCookie === true;
 
+  normalizedLoadOpts.sameDomainCookiesOnly = normalizedLoadOpts.sameDomainCookiesOnly === true;
+
   const uaChTrackLevels = ['none', 'default', 'full'];
   if (!uaChTrackLevels.includes(normalizedLoadOpts.uaChTrackLevel as UaChTrackLevel)) {
     normalizedLoadOpts.uaChTrackLevel = undefined;