+++ packages/analytics-js-common/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.9.5](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-common@3.9.4...@rudderstack/analytics-js-common@3.9.5) (2024-10-18)
+
+
+### Bug Fixes
+
+* add xpixel missing name mapping ([#1895](https://github.com/rudderlabs/rudder-sdk-js/issues/1895)) ([0800f36](https://github.com/rudderlabs/rudder-sdk-js/commit/0800f361b6a9b9866ba47d9a393cf7c2063b3815))
+
+## [3.9.4](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-common@3.9.3...@rudderstack/analytics-js-common@3.9.4) (2024-10-17)
+
+
+### Bug Fixes
+
+* renamed june destination display name ([324ba5d](https://github.com/rudderlabs/rudder-sdk-js/commit/324ba5d87184215cd1699fac66dff7ab79acc88a))
+
 ## [3.9.3](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-common@3.9.2...@rudderstack/analytics-js-common@3.9.3) (2024-09-27)