+++ packages/analytics-js/CHANGELOG.md
@@ -2,6 +2,25 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.7.19](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js@3.7.18...@rudderstack/analytics-js@3.7.19) (2024-10-18)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-cookies` updated to version `0.3.11`
+* `@rudderstack/analytics-js-common` updated to version `3.9.5`
+* `@rudderstack/analytics-js-plugins` updated to version `3.4.19`
+
+### Bug Fixes
+
+* harmless change in the core sdk package ([5c34b04](https://github.com/rudderlabs/rudder-sdk-js/commit/5c34b0453041a0dc1b8a7dc55eb0c884d16d0598))
+
+## [3.7.18](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js@3.7.17...@rudderstack/analytics-js@3.7.18) (2024-10-17)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-cookies` updated to version `0.3.10`
+* `@rudderstack/analytics-js-common` updated to version `3.9.4`
+* `@rudderstack/analytics-js-plugins` updated to version `3.4.18`
 ## [3.7.17](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js@3.7.16...@rudderstack/analytics-js@3.7.17) (2024-10-11)
 
 ### Dependency Updates